refactors sendeventbatch
